# Planner optimisations