### Description

This PR optimises that maths in `Planner::calculate_trapezoid_for_block()` and replaces `block->nominal_speed_sqr` with `block->nominal_speed`.

The latter is a simple matter: it eliminates multiple calls to `SQRT()` in exchange with a few calls to `sq()`.

The former involves manually inlining the functions `Planner::estimate_acceleration_distance()` and `Planner::intersection_distance()`. This makes it possible to re-use some calculations between these functions. The changes to the comments early in `planner.cpp` best explain the maths.

There is also a replacement of a division with a multiplication by an inverse (the compiler cleverly notices `(0.5 / accel)` is used more than once and re-uses the result). And a call to `FLOOR()` is removed because the argument was guaranteed to be a positive number which was going to be stored in an integer anyway, making the `FLOOR()` redundant. (The same could not be done with `CEIL()` on the previous line but does call into question whether `CEIL()` is actually required.)

On my AVR based printer, I profiled `Planner::_buffer_steps()` whilst drawing arcs. Before the changes:
- `G3 I-3 F10000` spent a total of 247ms of non-idle time in `Planner::_buffer_steps()` whilst drawing the arc.
- `G3 I-50 F10000` spent 1.43s.

After the changes:
- `G3 I-3 F10000` spent 217ms.
- `G3 I-50 F10000` spent 1.29s.

So that's well over 10% performance gain.

By `idle time` I mean time spent calling `idle()` whilst the planner waits for a free block in the queue.

### Requirements

Use the planner.

### Benefits

More spare CPU time for other things. Especially beneficial for AVR printers.

### Configurations

My config with which I tested: 
[config.zip](https://github.com/MarlinFirmware/Marlin/files/9087383/config.zip)

### Related Issues

Follow on from #24366.