@thinkyhead 2022-07-15
It seems good, and if it maintains the same motion profile in all cases, edge or otherwise, then I'm happy to merge it RSN.

I think the compiler was probably optimizing out the `FLOOR`, and if that is the case we might want to just keep it in there anyway, just in case some future or outside nerd decides to change those int types into float.