@thinkyhead 2022-07-15
Are we sure the change to the use of float types for the acceleration calculations will not impact the results and lead to disparity with the integer types alongside them? For the most part it seems like they should end up producing products in the same general range, but values > 0 and < 1 might have extra concern. Of course, such small values seem like they should be very rare or even impossible.