@thinkyhead 2022-07-15
For more optimization on processors with an FPU (and lots of Flash+RAM) we might consider using a type that is a `double` on those MCUs but a `float` on MCUs that don't, not just in the planner but throughout Marlin.