@ManuelMcLure 2022-07-15
> For more optimization on processors with an FPU (and lots of Flash+RAM) we might consider using a type that is a `double` on those MCUs but a `float` on MCUs that don't have FPU, not just in the planner but throughout Marlin.

The FPUs in Cortex-M4 processors (most of the FPU-enabled processors currently supported by Marlin) are single-precision, so in those cases we still want to use `float` instead of `double`.