@thinkyhead 2022-07-15
> The FPUs in Cortex-M4 processors … are single-precision…

Aha, well that simplifies things. And we don't really have much to gain from `double` anyway.