@tombrazier 2022-07-16
> I think the compiler was probably optimizing out the `FLOOR`, and if that is the case we might want to just keep it in there anyway, just in case some future or outside nerd decides to change those int types into float.

Changing FLOOR to truncation saved about 20ms with `G3 I-50 F10000` on my machine.

> Are we sure the change to the use of float types for the acceleration calculations will not impact the results and lead to disparity with the integer types alongside them?

The calculations all use floats and ints in the same places as before. Or at least that was the intention.